category api testleri yazıldı

+++ src/Services/Category/ContestPark.Category.API/Controllers/SubCategoryController.cs
@@ -2,7 +2,6 @@
 using ContestPark.Category.API.Infrastructure.Repositories.Category;
 using ContestPark.Category.API.Infrastructure.Repositories.FollowSubCategory;
 using ContestPark.Category.API.Infrastructure.Repositories.OpenCategory;
-using ContestPark.Category.API.Infrastructure.Repositories.Search;
 using ContestPark.Category.API.Model;
 using ContestPark.Category.API.Resources;
 using ContestPark.Core.CosmosDb.Models;
@@ -92,11 +91,10 @@ namespace ContestPark.Category.API.Controllers
         /// Alt kategori takip et
         /// </summary>
         /// <param name="subCategoryId">Alt kategori Id</param>
-        [HttpPost]
-        [Route("{subCategoryId}/Follow")]
+        [HttpPost("{subCategoryId}/Follow")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> Post(string subCategoryId)
+        public async Task<IActionResult> Post([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -104,9 +102,13 @@ namespace ContestPark.Category.API.Controllers
             if (_followSubCategoryRepository.IsSubCategoryFollowed(UserId, subCategoryId))// Kategoriyi daha önceden takip etmişmi
                 return BadRequest(CategoryResource.YouAreAlreadyFollowingThisCategory);
 
-            if (!(_categoryRepository.IsSubCategoryFree(subCategoryId) ||// kategori ücretsiz değil ise
-                _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId)))// veya kategorinin kilidi açık değilse
+            if (!(
+                _categoryRepository.IsSubCategoryFree(subCategoryId) ||// kategori ücretsiz değil ise
+                _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId)// veya kategorinin kilidi açık değilse
+                ))
+            {
                 return BadRequest(CategoryResource.ToBeAbleToFollowThisCategoryYouNeedToUnlockIt);
+            }
 
             bool isSuccess = await _followSubCategoryRepository.AddAsync(new FollowSubCategory
             {
@@ -138,10 +140,13 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}/UnFollow")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> Delete(string subCategoryId)
+        public async Task<IActionResult> Delete([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
+
+            if (!_followSubCategoryRepository.IsSubCategoryFollowed(UserId, subCategoryId))
+                return BadRequest(CategoryResource.YouMustFollowThisCategoryToDeactivateTheCategory);
 
             bool isSuccess = await _followSubCategoryRepository.DeleteAsync(UserId, subCategoryId);
             if (!isSuccess)
@@ -169,7 +174,7 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}/FollowStatus")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public IActionResult Get(string subCategoryId)
+        public IActionResult Get([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -188,7 +193,7 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}")]
         [ProducesResponseType(typeof(SubCategoryDetailModel), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public IActionResult GetDetail(string subCategoryId)
+        public IActionResult GetDetail([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -216,14 +221,17 @@ namespace ContestPark.Category.API.Controllers
         /// Alt kategori kilit açma
         /// </summary>
         /// <param name="subCategoryId">Alt kategori Id</param>
-        [HttpPost]
-        [Route("{subCategoryId}/unlock")]
+        [HttpPost("{subCategoryId}/unlock")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> UnLockSubCategory(string subCategoryId)
+        public async Task<IActionResult> UnLockSubCategory([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
+
+            bool isSubCategoryFree = _categoryRepository.IsSubCategoryFree(subCategoryId);
+            if (isSubCategoryFree)
+                return BadRequest(CategoryResource.YouCanNotUnlockTheFreeCategory);
 
             bool isSubCategoryOpen = _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId);
             if (isSubCategoryOpen)
